pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arabic_synth"
version = "0.1.0"
description = "Persona-based synthetic Arabic datasets generation pipeline"
authors = [{name = "AI PM Intern"}]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.7",
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "rapidfuzz>=3.10",
    "orjson>=3.10",
    "tqdm>=4.66",
    "tenacity>=8.2",
    "httpx>=0.27",
    "typer>=0.12",
    "pydantic-settings>=2.3",
    "langchain>=0.2.16",
    "sentence-transformers>=3.0",
    "jsonschema>=4.23",
    "python-dotenv>=1.0",
]

[project.scripts]
arabic-synth = "arabic_synth.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true 